Backend/pyproject.toml
[tool.poetry]
name = "prompt-proof-market-backend"
version = "0.1.0"
description = "FastAPI based backend for prompt-proof-market"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false  # 👈 move this up into [tool.poetry]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
python-dotenv = "^1.1.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.0"
requests = "^2.32.3"
atoma-sdk = "^0.1.3"
openai = "^1.75.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
